rename transfer helper to a different name
